+++ app/pb_kits/playbook/packs/samples.js
@@ -10,41 +10,6 @@ import WebpackerReact from 'webpacker-react'
 // ===========================================
 // Generated file from kit generator.
 
-import Dashboards from '../../../views/playbook/samples/dashboards/dashboards.jsx'
+import Dashboards from '../../../views/playbook/samples/dashboards/index.jsx'
 
 WebpackerReact.setup({ Dashboards })
-
-$(document).ready(() => {
-  $('.pb--codeCopy').show()
-  $('.compress').hide()
-
-  window.$(document).on('click', '.toggle-button', (e) => {
-    e.preventDefault()
-    $('.pb--codeCopy').toggleClass('close')
-  })
-
-  $('.expand').click(() => {
-    $('.sample-nav').hide()
-    $('.compress').show()
-  })
-
-  $('.compress').click(() => {
-    $('.sample-nav').show()
-    $('.compress').hide()
-  })
-
-  const setClipboard = (value) => {
-    var tempInput = document.createElement('textarea')
-    tempInput.style = 'position: absolute; left: -1000px; top: -1000px'
-    tempInput.value = value
-    document.body.appendChild(tempInput)
-    tempInput.select()
-    document.execCommand('copy')
-    document.body.removeChild(tempInput)
-  }
-
-  $('.copy-clipboard').click(() => {
-    var copyText = document.querySelector('.hiddenCodeforCopy').textContent
-    setClipboard(copyText)
-  })
-})
