+++ app/views/layouts/playbook/samples.html.erb
@@ -9,23 +9,6 @@
     <body>
         <div class="sample-layout">
             <%= yield %>
-        
-            <% if @type == "rails" %> 
-                <% contents = File.read("#{Playbook::Engine.root}/app/views/playbook/samples/dashboards/dashboards.html.erb") %>
-                <%= render template: "playbook/samples/dashboards/dashboards.html.erb", locals: { type: @type } %>
-            <% elsif @type == "react" %>
-                <% contents = File.read("#{Playbook::Engine.root}/app/views/playbook/samples/dashboards/dashboards.jsx") %>
-            <% end %>
-            <%= react_component("Dashboards") %>
-            <div class="pb--kit-show">
-                <div class="pb--doc light_ui">
-                    <div class="pb--codeCopy close" >
-                        <a class="pb--close-toggle copy-clipboard" href="#">Copy to Clipboard</a>
-                        <div class="hiddenCodeforCopy"><%= contents %></div>
-                        <%= raw rouge(contents, "erb") %>
-                    </div>
-                </div>
-            </div>
         </div>
     </body>
