+++ app/views/playbook/samples/dashboards/index.html.erb
