+++ app/views/playbook/samples/dashboards/index.jsx
