+++ app/views/playbook/samples/sample_show.html.erb
@@ -20,6 +20,20 @@
   </div>
 </div>
 
-
 <%= pb_rails("circle_icon_button", props:{ classname: "compress", variant: "link", icon: "compress"}) %>
-
+<% if @type == "rails" %> 
+    <% contents = File.read("#{Playbook::Engine.root}/app/views/playbook/samples/dashboards/dashboards.html.erb") %>
+    <%= render template: "playbook/samples/dashboards/dashboards.html.erb", locals: { type: @type } %>
+<% elsif @type == "react" %>
+    <% contents = File.read("#{Playbook::Engine.root}/app/views/playbook/samples/dashboards/dashboards.jsx") %>
+<% end %>
+<%= react_component("Dashboards") %>
+<div class="pb--kit-show">
+    <div class="pb--doc light_ui">
+        <div class="pb--codeCopy close" >
+            <a class="pb--close-toggle copy-clipboard" href="#">Copy to Clipboard</a>
+            <div class="hiddenCodeforCopy"><%= contents %></div>
+            <%= raw rouge(contents, "erb") %>
+        </div>
+    </div>
+</div>
